补充一些代码

+++ TTMS_Server/src/main/java/TTMS_Server/serviceimpl/StudioServiceImpl.java
@@ -1,9 +1,11 @@
-package TTMS_Server.serviceImpl;
+package TTMS_Server.serviceimpl;
 
 import TTMS_Server.dao.StudioDAO;
 import TTMS_Server.model.Studio;
+import TTMS_Server.service.ScheduleService;
 import TTMS_Server.service.SeatService;
 import TTMS_Server.service.StudioService;
+import TTMS_Server.service.TicketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +19,9 @@ public class StudioServiceImpl implements StudioService {
 
     @Autowired
     private SeatService seatService;
+
+    @Autowired
+    private ScheduleService scheduleService;
 
     //根据id获取演出厅信息
     @Override
@@ -50,7 +55,8 @@ public class StudioServiceImpl implements StudioService {
     //删除演出厅
     @Override
     public boolean deleteStudioById(Integer id){
-        if (studioDAO.selectStudioById(id) != null){
+        //当演出厅存在并且没有安排演出计划
+        if (studioDAO.selectStudioById(id) != null && scheduleService.selectScheduleByStudioId(id) == null){
             studioDAO.deleteStudioById(id);
             return true;
         }
